+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add option to create logger with text output format using kitlog.NewLogfmtLogger
+
 ## [1.0.0] - 2022-03-01
 
 ### Added